==> nsiegel/02-scrape_reviews.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"></ul></div>

import praw
import pandas as pd
import re

import requests
import time
import sys
import pdb

rdb = pd.read_parquet('redditwhiskeydatabase.parquet')

Client secrets for reddit are loaded here. You'll need to provide your own secrets.py which provides the following fields. You'll need to sign up for a reddit api to get this.:

- clientid
- clientsecret
- clientusername
- clientpassword

import secrets

def extract_reviews(dataframe,re_client_id,re_client_secret,\
                    re_username,re_pw):
        """Takes Reddit review submission ids, request \
        submissions from PRAW, search for the right body text,\
        and saves it to the dataframe""" 
        #timer/counter
        start_time = int(time.time())
        count = 0
        total = dataframe.shape[0]
        
        #get the Review URL addresses as a series
        rev_addr = dataframe['sub_id']

        #empty list of reviews
        reviews = []
        
        #initiate PRAW
        reddit = praw.Reddit(client_id= re_client_id,\
                     client_secret = re_client_secret,\
                     username = re_username, \
                             password = re_pw,user_agent= 'v1' )
        
        
        
        #iterate through the submission ids to extract review texts
        for s_id in rev_addr:
            review = ''
            try:
                sub = reddit.submission(id = s_id)
                s_author = sub.author.name
            except:
                reviews.append('error')
                continue
            #sometimes the reviews are in the 
            #submission text instead of comments
            #add any text in this field if not empty
            sub_text = sub.selftext
            if len(sub_text) > 0:
                review = review + sub_text + ' \n'

            # parse through the submission's comments 
            # and extract all of authors' comments
            for comment in sub.comments:
                try:
                    if comment.author.name == s_author:
                        review = review + comment.body + '\n'
                except:
                    continue
            
            reviews.append(review)
            
            #counter/timer
            end_time = int(time.time())
            min_spent = round(((end_time-start_time)/60),2)
            remaining_time = round((min_spent/max(count,1))*max((total-count),1),2)
            
            
            count_str = 'extracted ' + str(count) + '/' + str(total) + \
                    ' reviews in total ' + str(min_spent) + ' minutes' + \
                    ' estimated time remaining is ' + str(remaining_time) + ' minutes'
            sys.stdout.write("\r{}".format(count_str))
            count += 1
            
        dataframe['review'] = reviews
        #save to csv, also return the pandas dataframe
        dataframe.to_csv("db_reviews.csv")
        return dataframe
        

reviews = extract_reviews(rdb,clientid,clientsecret,\
                    clientusername,clientpassword)

reviews.shape

reviews.to_parquet('db_reviews.parquet')

reviews.tail()
