==> nsiegel/04-scrape_products.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"></ul></div>

import pandas as pd
import urllib.parse
import urllib.request
import xml.etree.ElementTree as ET
import io
from bs4 import BeautifulSoup
import time
import sys

onematch = pd.read_parquet('data/product_info140000.parquet')
onematch['categoryname']

onematch = pd.read_parquet('onematch.parquet')
product_ids = pd.read_parquet('product_ids.parquet')

product_info10000 = pd.read_parquet('product_info10000.parquet')

product_info10000.dropna().shape

Create a list of product ids we want to look up. We can then fill in all of the information about these product ids in a new table.

newIds = (pd.DataFrame(product_ids.ProductID.append(onematch.ProductID).unique(),columns=['ProductID'])
          .sort_values('ProductID')
          .reset_index(drop=True)
         )
newIds['Response'] = None
newIds['XML'] = None
newIds.head()

Make a list of ids to pull

newIds = list(product_ids.ProductID.append(onematch.ProductID).unique())
newIds.sort()

newIds[1:10]

product_info = scrape_all_info(newIds)

range(10,20)

product_info100000 = scrape_all_info2(range(100000))
product_info100000.to_parquet('product_info100000.parquet')

range(1,10)

for i in range(7,50): # 5,110
    maxrange = i * 10000
    minrange = (i-1) * 10000 + 1
    print('--')
    print(str(minrange) + ' to ' + str(maxrange))
    product_info_current = scrape_all_info2(range(minrange,maxrange))
    product_info_current.to_parquet('product_info'+str(maxrange)+'.parquet')
    

for i in range(5,110):
    maxrange = i * 10000
    minrange = (i-1) * 10000 + 1
    print('--')
    print(str(minrange) + ' to ' + str(maxrange))
    #product_info_current = scrape_all_info2(range(minrange,maxrange))
    #product_info_current.to_parquet('product_info'+str(maxrange)+'.parquet')
    

product_info100000.shape()

def scrape_all_info2(idlist):
    fieldlist = ['language', 'itemnumber', 'itemname', 'price', 'productsize', 'isdiscontinued', 'islimited',
                 'am', 'ambonusmiles', 'amexpiration', 'kosher', 
                 'lto', 'ltoexpiration', 'ltoregularprice', 'ltosavings', 
                 'releasedate', 'upcnumber', 'sccnumber', 'vao', 'vaodescription', 'vaoexpiration', 
                 'vqa', 'productquantity', 'producer', 'producingsubregion', 'producingregion', 'producingcountry', 
                 'alcoholpercentage', 'sugarcontent', 'tastingnotes', 'pairings', 
                 'itemdescription', 'liquortype', 'categoryname', 'stocktype', 'subcategoryname',
                 'sellingpackage', 'sweetnessdescriptor', 'winestyle', 'winevarietal', 
                 'stylebody', 'stylebody_fr', 'styleflavour', 'styleflavour_fr']

    productinfo = pd.DataFrame(columns=fieldlist)
    
    #timer/counter
    start_time = int(time.time())
    count = 0
    total = len(idlist)
    

    for id in idlist:
        #print(id)
        # Pull xml from lcbo api
        xmltest = pullXML(str(id))
        # Extract dict of values
        newitem = parse_item_xml(xmltest, fieldlist)

        # Add to our dataframe
        productinfo = productinfo.append(newitem,ignore_index=True)
        
        #counter/timer
        end_time = int(time.time())
        min_spent = round(((end_time-start_time)/60),2)
        remaining_time = round((min_spent/max(count,1))*max((total-count),1),2)

        count_str = 'extracted ' + str(count) + '/' + str(total) + \
                ' products in total ' + str(min_spent) + ' minutes' + \
                ' estimated time remaining is ' + str(remaining_time) + ' minutes'
        if count % 100 == 0:
            sys.stdout.write("\r{}".format(count_str))
        count += 1
        
        # sleep so we don't get blacklisted
        #time.sleep(0.5)
        
    #save to parquet, also return the pandas dataframe
    #productinfo.to_parquet("productinfo2.parquet")
    return productinfo

def scrape_all_info(idlist):
    fieldlist = ['language', 'itemnumber', 'itemname', 'price', 'productsize', 'isdiscontinued', 'islimited',
                 'am', 'ambonusmiles', 'amexpiration', 'kosher', 
                 'lto', 'ltoexpiration', 'ltoregularprice', 'ltosavings', 
                 'releasedate', 'upcnumber', 'sccnumber', 'vao', 'vaodescription', 'vaoexpiration', 
                 'vqa', 'productquantity', 'producer', 'producingsubregion', 'producingregion', 'producingcountry', 
                 'alcoholpercentage', 'sugarcontent', 'tastingnotes', 'pairings', 
                 'itemdescription', 'liquortype', 'categoryname', 'stocktype', 'subcategoryname',
                 'sellingpackage', 'sweetnessdescriptor', 'winestyle', 'winevarietal', 
                 'stylebody', 'stylebody_fr', 'styleflavour', 'styleflavour_fr']

    productinfo = pd.DataFrame(columns=fieldlist)
    
    #timer/counter
    start_time = int(time.time())
    count = 0
    total = len(idlist)

    for id in idlist:
        print(id)
        # Pull xml from lcbo api
        xmltest = pullXML(id)
        # Extract dict of values
        newitem = parse_item_xml(xmltest, fieldlist)

        # Add to our dataframe
        productinfo = productinfo.append(newitem,ignore_index=True)
        
        #counter/timer
        end_time = int(time.time())
        min_spent = round(((end_time-start_time)/60),2)
        remaining_time = round((min_spent/max(count,1))*max((total-count),1),2)

        count_str = 'extracted ' + str(count) + '/' + str(total) + \
                ' reviews in total ' + str(min_spent) + ' minutes' + \
                ' estimated time remaining is ' + str(remaining_time) + ' minutes'
        sys.stdout.write("\r{}".format(count_str))
        count += 1
        
        # sleep so we don't get blacklisted
        time.sleep(3)
        
    #save to parquet, also return the pandas dataframe
    productinfo.to_parquet("productinfo.parquet")
    return productinfo

def parse_item_xml(xmlstring, fieldlist):
    soup = BeautifulSoup(xmlstring)
    product = soup.find('products')
    
    newitem = {}
    
    for field in fieldlist:
        value = product.find(field)
        if value is not None:
            newitem[field] = value.text

    return newitem

def pullXML(ProductID):
    url = 'http://www.foodanddrink.ca/lcbo-webapp/productdetail.do?itemNumber=' + ProductID
    user_agent = 'Mozilla/5.0 (Windows NT 6.1; Win64; x64)'
    values = {'language' : 'en' }
    headers = { 'User-Agent' : user_agent }
    data = urllib.parse.urlencode(values)
    data = data.encode('ascii')
    req = urllib.request.Request(url, data, headers)
    try:
        with urllib.request.urlopen(req) as response:
           content = response.read()
        return content
        #return BeautifulSoup(content, 'xml')
    except :
        raise response
